UPDATE: Add testing rig for TypeScript SDK (#80)
It's far from ideal. Solution for now is to have two test runners:
- unit test runner directly in root to allow testing private
  functionalities that do not depend on WASM
- e2e tests that import node package and do test WASM functionality

Rollup builds a good, easy to import in node package, but test runners
work on unbundled files. So, WASM needs to be supported properly by
runners.

Jest 29.3.0 seems to support WASM [^1], but I was getting error:
  "Must use import to load ES Module"
regardless of the ts-jest settings. And to me, it appeared that import
is used.

[^1]: https://github.com/facebook/jest/pull/13505