UPDATE: Add evaluation checks on XLSX load in SDKs. Remove =TODAY fn (#114)
Idea is to fail fast without letting developers to work on broken
workbooks. If we notice any changes vs data stored in XLSX (XLSX stores
information about latest evaluation) or evaluation simply fails
(for example, because of unsupported functions) - we want to report it
as fast as possible.

Removes =TODAY function, so we do not support any volatile functions.